requirements.txt
asgiref==3.8.1
celery==5.4.0
Django==5.1.1
django-environ==0.11.2
djangorestframework==3.15.2
psycopg[binary]==3.2.9
redis==5.1.0
sqlparse==0.5.1
tzdata==2024.1
Pillow==10.4.0
requests==2.32.3
